Add liquibase.xsd-version as maven property
The XML parsing of liquibase change lists require the XSD referenced
in the change lists as a classpath resource.

Because OSGi only sees the classpath resources of the current bundle,
the schema needs to be included into the same bundle as the changelist
and copying all of the schemas when you only need one, seems
excessive.

Unfortunately the XSD version number doesn't include the patch version
so a separate maven property is needed.